+++ extension/manifest.common.json
@@ -33,7 +33,9 @@
         "file://*/*"
       ],
       "js": [
-        "youtube-iframe-api.js",
+        "vendors/youtube-iframe-api.js",
+        "vendors/jspdf.min.js",
+        "jspdf-fonts/msyh-normal.js",
         "browser-polyfill.js",
         "content.js"
       ],
